fixes merger for when file is merged twice
This happens in projects where an #include is normally guarded by #ifdef.
File A conditionally includes either B or C based on an #ifdef.
Both B and C include D, so D gets merged twice.
Previously, the offset to which a file view has been created is stored globally.
This means that the offset would be reused on a fresh include in the above situation.
Solved by storing offsets keyed by the NodeIndex of the parent file that is including and the NodeIndex of the file being included.