Fix crashes when run with eglot as a client
This just handles ConfigurationDidChange events in which the "mcglsl"
key wasn't set without crashing, by ignoring them; if the
configuration changed, but none of our configuration changed, there's
nothing we need to update.

Obviously there is still a lot that can go wrong here, but this allows
running the language server with https://github.com/joaotavora/eglot,
as that language server sends a ConfigurationDidChange on startup if
any configuration is set (including for other language servers).